move autocomplete search to commons module
