+++ src/modules/commons/components/autocomplete/index.js
@@ -1,0 +1,4 @@
+import { AutocompleteSearch } from './AutocompleteSearch';
+if (!window.customElements.get(AutocompleteSearch.tag)) {
+	window.customElements.define(AutocompleteSearch.tag, AutocompleteSearch);
+}
