+++ src/modules/commons/index.js
@@ -2,3 +2,4 @@ import './components/button';
 import './components/toggle';
 import './components/popup';
 import './components/icon';
+import './components/autocomplete';
